receipt-reporter/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
jinja2==3.1.4
easyocr==1.7.1
openpyxl==3.1.5
python-docx==1.1.2
reportlab==4.2.2
Pillow==10.4.0
